+++ extensions/rednote-viewer/src/api.ts
@@ -1,11 +1,11 @@
-import fs from "node:fs";
+import fs from "node:fs/promises";
 import path from "node:path";
 import got from "got";
 import { UserInfo, BaseResponse, NoteRequest, NoteInfo, SearchRequest, HomeFeedRequest } from "./types.js";
 import { API_URL } from "./constants.js";
 
 /**
- * Get my info(Verify cookie is valid)
+ * Get my info (verify cookie is valid)
  * @param cookie cookie
  * @returns User info
  */
@@ -54,7 +54,7 @@ export const downloadAssetApi = async (url: string, dir: string) => {
   const file = await got(url).buffer();
   const filename = url.split("/").pop() || Date.now().toString();
   const destinationFile = path.join(dir, `${filename}`);
-  await fs.promises.writeFile(destinationFile, file);
+  await fs.writeFile(destinationFile, file);
 };
 
 /**
@@ -86,5 +86,8 @@ export const searchPostApi = async (data: SearchRequest, headers: Record<string,
 export const getHomeFeedApi = async (data: HomeFeedRequest, headers: Record<string, string>) => {
   const url = `${API_URL}/sns/web/v1/homefeed`;
   const response = await got.post(url, { headers, json: data }).json<BaseResponse<NoteInfo>>();
+  if (response.code !== 0) {
+    throw new Error(response.message);
+  }
   return response.data;
 };